feat: Add forceUpdate method to ComputedStore
This commit adds a new method, `forceUpdate`, to the `ComputedStore` interface. The `forceUpdate` method allows for manually triggering a re-computation of the computed value, by calling the `computedValueFn` with the previous states. This can be useful in scenarios where the computed value needs to be updated outside of the normal dependency tracking mechanism.

Note: This message does not include any meta information such as issue references, tags, or author names.